+++ e_ood_django/views.py
@@ -1,13 +1,14 @@
 from django.core.urlresolvers import reverse
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
+from django.utils.timezone import now
 from django.views.generic.edit import FormView, View
 
 from .forms import DBImportForm
 from .models import Package, PackageDBAccess
 
 
-class ExportDBView(View):
+class AnyExportDBView(View):
 
     @staticmethod
     def format_value(value):
@@ -16,8 +17,7 @@ class ExportDBView(View):
         else:
             return "'{}'".format(value)
 
-    def get(self, request, uuid):
-        get_object_or_404(PackageDBAccess, uuid=uuid)
+    def get_yaml(self):
         d = Package.export()
         # This sucks, but I don't see a clean way to output this as desired
         # in a clean manner.  The best I found was
@@ -29,7 +29,34 @@ class ExportDBView(View):
             for k, value in d[name].items():
                 y.append('  ' + k + ': ' + self.format_value(value))
         y = '\n'.join(y) + '\n'
+        return y
+
+
+class ExportDBView(AnyExportDBView):
+    """
+    Random clients can use this view as long as they have the magic
+    UUID.
+    """
+
+    def get(self, request, uuid):
+        get_object_or_404(PackageDBAccess, uuid=uuid)  # authorize client
+        y = self.get_yaml()
         return HttpResponse(y)
+
+
+class AdminExportDBView(AnyExportDBView):
+    """
+    Admin users can use this view from the button in admin.  The URL
+    spec wraps this with a staff-user requirement.
+    """
+
+    def get(self, request):
+        y = self.get_yaml()
+        # no registered MIME type for YAML as of early 2018 AFAICT
+        response = HttpResponse(content=y, content_type='text/x-yaml')
+        suggested_filename = now().strftime('db-%Y-%m-%d.yaml')
+        response['Content-Disposition'] = 'attachment; filename=%s' % suggested_filename
+        return response
 
 
 class ImportDBView(FormView):