+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes and migration requirements
 
+## Version 0.0.4
+
+* Add admin interface for exporting the package db.
+* Improve field names and make other minor improvements in admin.
+
 ## Version 0.0.3
 
 * Trivial admin improvements